Remove Item Price and add Item Measurement fields
This commit removes the "Item Price" field from the
`CreateItem.vue` and `EditItem.vue` components, replacing
it with an "Item Measurement" field. The `input` reactive
object has been updated to remove the `price` property
and include a `measurement` property initialized to an
empty string. Additionally, the item price display in
`item.vue` has been replaced with the item measurement.